refactor(examples): split esbuild config into focused modules
