+++ examples/scripts/esbuild/build-tasks.ts
@@ -1,0 +1,61 @@
+/**
+ * Copyright 2023-present DreamNum Co., Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import esbuild from 'esbuild';
+import cleanPlugin from 'esbuild-plugin-clean';
+
+export function monacoBuildTask() {
+    const monacoEditorEntryPoints = [
+        'vs/language/json/json.worker.js',
+        'vs/language/css/css.worker.js',
+        'vs/language/html/html.worker.js',
+        'vs/language/typescript/ts.worker.js',
+        'vs/editor/editor.worker.js',
+    ];
+
+    return esbuild.build({
+        entryPoints: monacoEditorEntryPoints.map((entry) => `./node_modules/monaco-editor/esm/${entry}`),
+        bundle: true,
+        color: true,
+        target: 'chrome70',
+        format: 'iife',
+        outbase: './node_modules/monaco-editor/esm/',
+        outdir: './local',
+        plugins: [
+            cleanPlugin({
+                patterns: ['./local'],
+            }),
+        ],
+    });
+}
+
+export function nodeBuildTask() {
+    return esbuild.build({
+        bundle: true,
+        color: true,
+        minify: false,
+        target: 'chrome70',
+        entryPoints: [
+            './src/node/cases/basic.ts',
+            './src/node/sdk/worker.ts',
+        ],
+        platform: 'node',
+        outdir: './dist',
+        define: {
+            'process.env.NODE_ENV': '"production"',
+        },
+    });
+}
